21-07-complite

+++ src/main/java/ru/kata/spring/boot_security/demo/Repository/RoleRepository.java
@@ -4,12 +4,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.kata.spring.boot_security.demo.entity.Role;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
+
     List<Role> findAllByNameIn(List<String> names);
     Role findByName(String name);
-
 }
